refactor: Change namespace on correct
